feat: add text label for name
